backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
numpy>=1.24.3
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.1.1
langchain>=0.1.0
langchain-community>=0.1.0
google-generativeai>=0.4.1